# Conflicts:
#	uniffi_bindgen/src/bindings/kotlin/templates/CallbackInterfaceImpl.kt
#	uniffi_bindgen/src/bindings/kotlin/templates/CallbackInterfaceRuntime.kt
#	uniffi_bindgen/src/bindings/kotlin/templates/ObjectCleanerHelperAndroid.kt

+++ uniffi_bindgen/src/bindings/ruby/templates/RustBufferTemplate.rb
@@ -1,6 +1,6 @@
 class RustBuffer < FFI::Struct
-  layout :capacity, :int32,
-         :len,      :int32,
+  layout :capacity, :uint64,
+         :len,      :uint64,
          :data,     :pointer
 
   def self.alloc(size)
@@ -181,7 +181,7 @@ class RustBuffer < FFI::Struct
 
   {% when Type::Optional { inner_type } -%}
   # The Optional<T> type for {{ canonical_name(inner_type) }}.
-  
+
   def self.check_lower_{{ canonical_type_name }}(v)
     if not v.nil?
       {{ "v"|check_lower_rb(inner_type.borrow()) }}
